Update to support CREV with CheckRevisionD1.mpq
It doesn't extract the data from the files on its own, but lets you put in the extracted values (which are the same every time unless there is an update) in the settings.ini file.

CheckRevision.mpq requires the "Version" field to be filled out.
CheckRevisionD1.mpq requires both "Version" and "Cert".
Cert should be the hex representation of the EXE's certificate public key, with no spaces.